Extract some config
